requirements.txt
# Core dependencies
numpy>=1.21.0
opencv-python>=4.5.0
PyYAML>=6.0
Pillow>=8.0.0

# Machine Learning
onnxruntime>=1.12.0
torch>=1.12.0
torchvision>=0.13.0

# Tracking
scipy>=1.7.0
scikit-learn>=1.0.0

# Jetson Nano specific (optional)
# onnxruntime-gpu  # For TensorRT support on Jetson

# Development dependencies
pytest>=6.0.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0